rename stark-felt -> starknet-types-core && Implement curve mod
