+++ Cargo.toml
@@ -1,7 +1,5 @@
 [workspace]
 members = [
-    "crates/stark-felt",
-    "crates/stark-curve",
     "crates/starknet-types-core",
     "crates/starknet-types-rpc",
 ]
