+++ crates/starknet-types-core/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "stark-felt"
+name = "starknet-types-core"
 version = "0.0.3"
 edition = "2021"
 license = "MIT"
@@ -13,7 +13,8 @@ readme = "README.md"
 [dependencies]
 bitvec = { version = "1.0.1", default-features = false }
 serde = { version = "1.0.163", optional = true, default-features = false }
-lambdaworks-math = { git = "https://github.com/Lambdaclass/lambdaworks.git" ,rev = "23154c1b2b7c1cb1620e40ee53fbb446a2923923", default_features = false }
+lambdaworks-math = { path = "../../../lambdaworks/math", default-features = false}
+
 
 arbitrary = { version = "1.3.0", optional = true, default-features = false }
 num-traits = { version = "0.2.16", default-features = false }
