+++ crates/starknet-types-core/src/felt.rs
@@ -1,12 +1,7 @@
-#![cfg_attr(not(feature = "std"), no_std)]
-
 use core::ops::{Add, Neg};
 
 use bitvec::array::BitArray;
 use num_traits::{FromPrimitive, ToPrimitive, Zero};
-
-#[cfg(test)]
-mod arbitrary_proptest;
 
 #[cfg(target_pointer_width = "64")]
 pub type BitArrayStore = [u64; 4];
@@ -33,7 +28,7 @@ use arbitrary::{self, Arbitrary, Unstructured};
 
 /// Definition of the Field Element type.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct Felt(pub FieldElement<Stark252PrimeField>);
+pub struct Felt(pub(crate) FieldElement<Stark252PrimeField>);
 
 /// A non-zero [Felt].
 pub struct NonZeroFelt(FieldElement<Stark252PrimeField>);
@@ -830,7 +825,7 @@ mod errors {
 mod test {
     use super::alloc::{format, string::String, vec::Vec};
     use super::*;
-    use crate::arbitrary_proptest::nonzero_felt;
+    use crate::felt_arbitrary::nonzero_felt;
     use core::ops::Shl;
     use proptest::prelude::*;
