+++ crates/starknet-types-core/src/lib.rs
@@ -1,0 +1,6 @@
+#![cfg_attr(not(feature = "std"), no_std)]
+pub mod curve;
+pub mod felt;
+
+#[cfg(test)]
+mod felt_arbitrary;
